+++ tests/page-objects/Browser/Confirmations/FooterActions.ts
@@ -16,7 +16,7 @@ class FooterActions {
     return Matchers.getElementByID(ConfirmationFooterSelectorIDs.CANCEL_BUTTON);
   }
 
-  async waitForConfirmButton(timeout = 45_000): Promise<void> {
+  async waitForConfirmButton(timeout = 30_000): Promise<void> {
     await Assertions.expectElementToExist(this.confirmButton, {
       timeout,
       description: 'confirm-button',
@@ -27,21 +27,14 @@ class FooterActions {
    * Wait until confirm-button leaves the hierarchy (does not use isDisplayed).
    * BottomSheet children often report isDisplayed=false while still present.
    */
-  async waitForConfirmButtonGone(timeout = 45_000): Promise<void> {
+  async waitForConfirmButtonGone(timeout = 30_000): Promise<void> {
     await Utilities.waitForElementToDisappear(this.confirmButton, timeout);
   }
 
   async tapConfirmButton(timeout?: number): Promise<void> {
     await ToastModal.waitForToastToDismiss({ appearTimeout: 2_000 });
 
-    // assets-controller@15.0.0's RpcFallbackMiddleware now also runs on
-    // regular background polls (previously forced-refresh only), adding a
-    // real RPC round trip to the local test chain's node on every poll tick
-    // since it's never in the (mocked) Accounts API's supportedNetworks list.
-    // That competes with TransactionController's own gas-estimation RPC
-    // calls on the same node, so gas fee resolution — which gates
-    // `checkEnabled` below — needs more headroom than before.
-    const readyTimeout = timeout ?? 45_000;
+    const readyTimeout = timeout ?? 30_000;
     await Gestures.waitAndTap(this.confirmButton, {
       elemDescription: 'Confirm button',
       timeout: readyTimeout,
