+++ tests/smoke-appium/confirmations/transactions/contract-interaction.spec.ts
@@ -24,7 +24,6 @@ import { setupRemoteFeatureFlagsMock } from '../../../api-mocking/helpers/remote
 import { confirmationFeatureFlags } from '../../../api-mocking/mock-responses/feature-flags-mocks.js';
 import { LocalNode, LocalNodeType } from '../../../framework/types.js';
 import { AnvilManager } from '../../../seeder/anvil-manager.js';
-import { ANVIL_LOCAL_ETH_HOLDING } from '../helpers/anvil-local-eth-holding.js';
 
 const NFT_CONTRACT = SMART_CONTRACTS.NFTS;
 
@@ -45,7 +44,6 @@ function buildContractInteractionFixture({
       nickname: 'Local RPC',
       ticker: 'ETH',
     })
-    .withTokenHoldings([ANVIL_LOCAL_ETH_HOLDING])
     .withPermissionControllerConnectedToTestDapp(buildPermissions(['0x539']))
     .build();
