+++ tests/smoke-appium/confirmations/transactions/gas-fee-tokens-eip-7702-sponsored.spec.ts
@@ -13,7 +13,6 @@ import {
 } from '../../../flows/confirmations.flow.js';
 import { withFixtures } from '../../../framework/fixtures/FixtureHelper.js';
 import { AnvilManager, Hardfork } from '../../../seeder/anvil-manager.js';
-import { ANVIL_LOCAL_ETH_HOLDING } from '../helpers/anvil-local-eth-holding.js';
 import {
   setupMockRequest,
   setupMockPostRequest,
@@ -182,7 +181,6 @@ const createFixture = ({ localNodes }: { localNodes?: LocalNode[] }) => {
       nickname: 'Local RPC',
       ticker: 'ETH',
     })
-    .withTokenHoldings([ANVIL_LOCAL_ETH_HOLDING])
     .withDisabledSmartTransactions()
     .build();
 };
