+++ tests/smoke-appium/confirmations/transactions/token-approve/approve.helpers.ts
@@ -12,7 +12,6 @@ import { setupRemoteFeatureFlagsMock } from '../../../../api-mocking/helpers/rem
 import { confirmationFeatureFlags } from '../../../../api-mocking/mock-responses/feature-flags-mocks.js';
 import { LocalNode } from '../../../../framework/types.js';
 import { AnvilManager } from '../../../../seeder/anvil-manager.js';
-import { ANVIL_LOCAL_ETH_HOLDING } from '../../helpers/anvil-local-eth-holding.js';
 
 export const ERC_20_CONTRACT = SMART_CONTRACTS.HST;
 export const ERC_721_CONTRACT = SMART_CONTRACTS.NFTS;
@@ -35,7 +34,6 @@ export function buildApproveFixture({
       nickname: 'Local RPC',
       ticker: 'ETH',
     })
-    .withTokenHoldings([ANVIL_LOCAL_ETH_HOLDING])
     .withPermissionControllerConnectedToTestDapp(buildPermissions(['0x539']))
     .build();
