+++ tests/smoke-appium/multichain/wallet-invokeMethod-eip5792.spec.ts
@@ -18,12 +18,11 @@ import { Mockttp } from 'mockttp';
 import { setupRemoteFeatureFlagsMock } from '../../api-mocking/helpers/remoteFeatureFlagsHelper.js';
 import { remoteFeatureEip7702 } from '../../api-mocking/mock-responses/feature-flags-mocks.js';
 import { isHexString } from '@metamask/utils';
-import { ANVIL_LOCAL_ETH_HOLDING } from '../confirmations/helpers/anvil-local-eth-holding.js';
 
 function buildAnvilLocalEthFixture() {
   return new FixtureBuilder()
     .withDefaultFixture()
-    .withTokenHoldings([ANVIL_LOCAL_ETH_HOLDING])
+    .withAnvilLocalEthBalance()
     .build();
 }
